django_api/requirements.txt
asgiref==3.2.3
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==3.0.2
django-braces==1.14.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-oauth-toolkit==1.2.0
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.11.0
djangorestframework-jsonapi==3.0.0
drf-yasg==1.17.0
gunicorn==20.0.4
httplib2==0.17.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.1
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pyparsing==2.4.6
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
ruamel.yaml==0.16.6
ruamel.yaml.clib==0.2.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
